Users/qiuyufan/AI nurse/nurse_ally/requirements.txt
flask==2.2.3
flask-session==0.4.0
werkzeug==2.2.3
openai==0.27.0
python-dotenv==1.0.0
requests==2.28.2
geopy==2.3.0
pytz==2022.7.1
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.2
click==8.1.3
uuid==1.30